fix(navlink-updater.js): never add icon if update_at data of page is not published
Regardless read_at data for the page exist on user's local storage, if website didn't
publish update_at data for that page, don't add any icon to it. This case is possible when
user marks a page as read but then website owner decides to not use mark as read feature
anymore on that page.